% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{plot.cadist}
\alias{plot.cadist}
\title{Distribution plotting}
\usage{
\method{plot}{cadist}(x, var, main = NULL, sub = NULL, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{x}{Output of \code{\link{ca}} command with
\code{interest = "dist"}.}

\item{var}{Name of variable for plotting}

\item{main}{Main title of the plot. Defualt is NULL.}

\item{sub}{Sub title of the plot. Default is NULL.}

\item{xlab}{x-axis label. Default is NULL.}

\item{ylab}{y-axis label. Default is NULL.}

\item{...}{graphics parameters to be passed to the plotting
routines.}
}
\description{
Plots distributions and joint uniform confidence bands of variables in
interest from \code{\link{ca}} command.
}
\examples{
data("mortgage")
fm <- deny ~ black + p_irat
t <- c(rep(1, 2), rep(0, 14)) # Specify variables in interest
# Drawing Distributions
CAgraph <- ca(fm=fm, data = mortgage, var = "black", method = "logit",
              t = t, interest = "dist")
plot(x = CAgraph, var = "p_irat")

}
