% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{summary.camoment}
\alias{summary.camoment}
\title{Return the output of \code{\link{ca}} function.}
\usage{
\method{summary}{camoment}(object, ...)
}
\arguments{
\item{object}{Output of \code{\link{ca}} command with
\code{interest = "moments"}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Return the output of \code{\link{ca}} function.
}
