% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop.R
\name{plot.subpop}
\alias{plot.subpop}
\title{Plot 2-dimensional projections of variables in interest.}
\usage{
\method{plot}{subpop}(x, varx, vary, main = NULL, sub = NULL,
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{Output of \code{\link{subpop}} command.}

\item{varx}{The name of the variable to be plotted on the x-axis.}

\item{vary}{The name of the variable name to be plotted on the
y-axis.}

\item{main}{Main title of the plot. Default is \code{NULL}.}

\item{sub}{Sub title of the plot. Default is NULL.}

\item{xlab}{x-axis label. Default is \code{NULL}.}

\item{ylab}{y-axis label. Default is \code{NULL}.}

\item{...}{Graphics parameters to be passed to the plotting
routines.}
}
\description{
Takes output from \code{\link{subpop}} command as inputs and plots
2-dimensional projection plots of two specified variables. If a
variable in interest is of type factor, then the user must put it on
the y-axis.The range of x-axis is set to be the range of variable on the
x-coordinate. If the variable on the y-coordinate is a factor, range of
y-axis is set to be the factor level; otherwise it is set to be the range of
the variable.
}
\examples{
data("mortgage")
fm <- deny ~ black + p_irat + hse_inc
result <- subpop(fm = fm, data = mortgage, var = "black", method = "logit")
plot(x = result, varx = mortgage$p_irat, vary = mortgage$ccred)

}
