% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wage2015}
\alias{wage2015}
\title{Wage Data}
\format{Consists of white, non-hispanic individuals aging from 25 to 64 and working more than
35 hours per week during at least 50 weeks of the year. Excludes self-employed, individuals
living in group quarters; individuals in the military, agricultural or private household sectors;
individuals with inconsistent reports on earnings and employment status; individuals with allocated
or missing information in any of the variables used in the analysis; individuals with hourly wage
rate below $3. Contains 32,523 workers including 18,137 men and 14,382 women.}
\source{
U.S. March Supplement of the Current Population Survey (CPS) in 2015.
}
\usage{
wage2015
}
\description{
Wage Data
}
\keyword{datasets}
